+++ INTROOT/SpirouDRS/data/constants/recipe_control.txt
@@ -48,10 +48,6 @@ cal_HC_E2DS_spirou, 1, e2dsff_AB, None, None, None, AB, None, hc, REDUC, HC, -6,
 cal_HC_E2DS_spirou, 1, e2dsff_A, None, None, None, A, None, hc, REDUC, HC, -6, 1
 cal_HC_E2DS_spirou, 1, e2dsff_B, None, None, None, B, None, hc, REDUC, HC, -6, 1
 cal_HC_E2DS_spirou, 1, e2dsff_C, None, None, None, C, None, hc, REDUC, HC, -6, 1
-cal_HC_E2DS_spirou, 1, e2dsff_AB, None, None, None, AB, HCONE_HCONE, hc, REDUC, HC, -6, 1
-cal_HC_E2DS_spirou, 1, e2dsff_A, None, None, None, A, HCONE_HCONE, hc, REDUC, HC, -6, 1
-cal_HC_E2DS_spirou, 1, e2dsff_B, None, None, None, B, HCONE_HCONE, hc, REDUC, HC, -6, 1
-cal_HC_E2DS_spirou, 1, e2dsff_C, None, None, None, C, HCONE_HCONE, hc, REDUC, HC, -6, 1
 cal_WAVE_E2DS_spirou, 1, e2ds_AB, None, None, None, AB, HCONE_HCONE, hc, REDUC, WAVE, -7, 1
 cal_WAVE_E2DS_spirou, 1, e2ds_A, None, None, None, A, HCONE_HCONE, hc, REDUC, WAVE, -7, 1
 cal_WAVE_E2DS_spirou, 1, e2ds_B, None, None, None, B, HCONE_HCONE, hc, REDUC, WAVE, -7, 1
@@ -80,14 +76,6 @@ cal_WAVE_NEW_E2DS_spirou, 1, e2dsff_AB, None, None, None, AB, HCONE_HCONE, hc, R
 cal_WAVE_NEW_E2DS_spirou, 1, e2dsff_A, None, None, None, A, HCONE_HCONE, hc, REDUC, WAVE, -7, 1
 cal_WAVE_NEW_E2DS_spirou, 1, e2dsff_B, None, None, None, B, HCONE_HCONE, hc, REDUC, WAVE, -7, 1
 cal_WAVE_NEW_E2DS_spirou, 1, e2dsff_C, None, None, None, C, HCONE_HCONE, hc, REDUC, WAVE, -7, 1
-cal_WAVE_NEW_E2DS_spirou, 1, e2dsff_AB, None, None, None, AB, HCONE_FP, hc, REDUC, WAVE, -7, 1
-cal_WAVE_NEW_E2DS_spirou, 1, e2dsff_A, None, None, None, A, HCONE_FP, hc, REDUC, WAVE, -7, 1
-cal_WAVE_NEW_E2DS_spirou, 1, e2dsff_B, None, None, None, B, HCONE_FP, hc, REDUC, WAVE, -7, 1
-cal_WAVE_NEW_E2DS_spirou, 1, e2dsff_C, None, None, None, C, HCONE_FP, hc, REDUC, WAVE, -7, 1
-cal_WAVE_NEW_E2DS_spirou, 1, e2dsff_AB, None, None, None, AB, FP_HCONE, hc, REDUC, WAVE, -7, 1
-cal_WAVE_NEW_E2DS_spirou, 1, e2dsff_A, None, None, None, A, FP_HCONE, hc, REDUC, WAVE, -7, 1
-cal_WAVE_NEW_E2DS_spirou, 1, e2dsff_B, None, None, None, B, FP_HCONE, hc, REDUC, WAVE, -7, 1
-cal_WAVE_NEW_E2DS_spirou, 1, e2dsff_C, None, None, None, C, FP_HCONE, hc, REDUC, WAVE, -7, 1
 cal_WAVE_NEW_E2DS_spirou, 2, e2dsff_AB, None, None, None, AB, FP_FP, fp, REDUC, WAVE, -7, 1
 cal_WAVE_NEW_E2DS_spirou, 2, e2dsff_A, None, None, None, A, FP_FP, fp, REDUC, WAVE, -7, 1
 cal_WAVE_NEW_E2DS_spirou, 2, e2dsff_B, None, None, None, B, FP_FP, fp, REDUC, WAVE, -7, 1
@@ -144,16 +132,6 @@ obj_mk_tell_template, 1, e2dsff_C, None, None, None, C, None, None, REDUC, MKTEL
 cal_exposure_meter, 1, o, None, None, None, None, None, None, RAW, EXTRACT, -12, 1
 cal_wave_mapper, 1, o, None, None, None, None, None, None, RAW, EXTRACT, -13, 1
 pol_spirou, 1, e2ds, None, None, None, None, None, None, REDUC, POLARIZATION, -14, 4
-visu_E2DS_spirou, 1, e2ds_AB, None, None, None, AB, None, None, REDUC, VISUE2DS -6, 1
-visu_E2DS_spirou, 1, e2ds_A, None, None, None, A, None, None, REDUC, VISUE2DS, -6, 1
-visu_E2DS_spirou, 1, e2ds_B, None, None, None, B, None, None, REDUC, VISUE2DS, -6, 1
-visu_E2DS_spirou, 1, e2ds_C, None, None, None, C, None, None, REDUC, MVISUE2DS, -6, 1
-visu_E2DS_spirou, 1, e2dsff_AB, None, None, None, AB, None, None, REDUC, VISUE2DS, -6, 1
-visu_E2DS_spirou, 1, e2dsff_A, None, None, None, A, None, None, REDUC, VISUE2DS, -6, 1
-visu_E2DS_spirou, 1, e2dsff_B, None, None, None, B, None, None, REDUC, VISUE2DS, -6, 1
-visu_E2DS_spirou, 1, e2dsff_C, None, None, None, C, None, None, REDUC, VISUE2DS, -6, 1
-
-
 #
 # Add extra DRS files which aren't currently used in recipes
 #
