+++ INTROOT/SpirouDRS/spirouTHORCA/spirouWAVE.py
@@ -395,9 +395,10 @@ def fp_wavelength_sol_new(p, loc):
     :param loc: parameter dictionary, ParamDict containing data
         Must contain at least:
             FPDATA: the FP e2ds data
-            LITTROW_EXTRAP_SOL_1: the wavelength solution derived from the HC
+            WAVE: the initial wavelength solution
+            LL_OUT_2: the wavelength solution derived from the HC
                                   and Littrow-constrained
-            LL_PARAM_1: the parameters of the wavelength solution
+            LL_PARAM_2: the parameters of the wavelength solution
             ALL_LINES_2: list of numpy arrays, length = number of orders
                        each numpy array contains gaussian parameters
                        for each found line in that order
@@ -427,7 +428,7 @@ def fp_wavelength_sol_new(p, loc):
 
     # find FP lines
     loc = find_fp_lines_new(p, loc)
-    all_lines_2 = loc['ALL_LINES_2']
+
     # set up storage
     llpos_all, xxpos_all, ampl_all = [], [], []
     m_fp_all, weight_bl_all, order_rec_all, dopd_all = [], [], [], []
@@ -439,8 +440,7 @@ def fp_wavelength_sol_new(p, loc):
         # select the lines in the order
         gg = loc['ORDPEAK'] == order_num
         # store the initial wavelengths of the lines
-        floc['llpos'] = np.polyval(loc['LITTROW_EXTRAP_PARAM_1'][order_num][::-1],
-                                   loc['XPEAK'][gg])
+        floc['llpos'] = np.polyval(loc['LL_PARAM_2'][order_num][::-1], loc['XPEAK'][gg])
         # store the pixel positions of the lines
         floc['xxpos'] = loc['XPEAK'][gg]
         # get the median pixel difference between successive lines (to check for gaps)
@@ -490,24 +490,6 @@ def fp_wavelength_sol_new(p, loc):
             if floc['llpos'][-1] > ll_prev[0]:
                 # find closest peak in overlap and get its m value
                 ind = np.abs(ll_prev - floc['llpos'][-1]).argmin()
-                # the peak matching the reddest may not always be found!!
-                # define maximum permitted difference
-                llpos_diff_med = np.median(floc['llpos'][1:]-floc['llpos'][:-1])
-                print(llpos_diff_med)
-                print(abs(ll_prev[ind] - floc['llpos'][-1]))
-                # check if the difference is over the limit
-                if abs(ll_prev[ind] - floc['llpos'][-1]) > 1.5*llpos_diff_med:
-                    print('overlap line not matched')
-                    ll_diff = ll_prev[ind] - floc['llpos'][-1]
-                    ind2 = -2
-                    # loop over next reddest peak until they match
-                    while ll_diff > 1.5*llpos_diff_med:
-                        # check there is still overlap
-                        if floc['llpos'][ind2] > ll_prev[0]:
-                            ind = np.abs(ll_prev - floc['llpos'][ind2]).argmin()
-                            ll_diff = ll_prev[ind] - floc['llpos'][ind2]
-                            ind2 -= 1
-                        else: break
                 m_match = m_prev[ind]
                 # save previous mpeak calculated
                 m_init = mpeak[-1]
@@ -559,7 +541,7 @@ def fp_wavelength_sol_new(p, loc):
     # update line wavelengths using the new cavity width fit
     newll = (dopd0 + cfit * 1000.) / m_fp_all
     # insert fp lines into all_lines2 (at the correct positions)
-    all_lines_2 = insert_fp_lines(p, newll, llpos_all, all_lines_2,
+    all_lines_2 = insert_fp_lines(p, newll, llpos_all, loc['ALL_LINES_2'],
                                   order_rec_all, xxpos_all, ampl_all)
 
     # add to loc
@@ -766,8 +748,6 @@ def find_hc_gauss_peaks(p, loc):
     WLOG('', p['LOG_OPT'], wmsg.format(os.path.basename(ini_table_name)))
     spirouImage.WriteTable(ini_table, ini_table_name, fmt='ascii.rst')
 
-    sPlt.wave_ea_plot_allorder_hcguess(p, loc)
-
     # return loc
     return loc
 
@@ -820,7 +800,6 @@ def fit_gaussian_triplets(p , loc):
     # TODO: Move loop out of function
     # set up storage
     wave_catalog = []
-    amp_catalog = []
     wave_map2 = np.zeros((nbo, nbpix))
     sig = np.nan
 
@@ -835,8 +814,6 @@ def fit_gaussian_triplets(p , loc):
         gauss_rms_dev = np.array(loc['GAUSS_RMS_DEV_INI'])
         ew = np.array(loc['EW_INI'])
         peak = np.array(loc['PEAK_INI'])
-        # get peak again for saving (to make sure nothing goes wrong in selection)
-        peak2 = np.array(loc['PEAK_INI'])
         # --------------------------------------------------------------
         # find the brightest lines for each order, only those lines will
         #     be used to derive the first estimates of the per-order fit
@@ -844,7 +821,7 @@ def fit_gaussian_triplets(p , loc):
         brightest_lines = np.zeros(len(xgau), dtype=bool)
         # loop around order
         for order_num in set(orders):
-            # find all order_nums that belong to this order
+            # find all order_nums that below to this order
             good = orders == order_num
             # get the peaks for this order
             order_peaks = peak[good]
@@ -864,7 +841,7 @@ def fit_gaussian_triplets(p , loc):
         ini_wave_sol = np.zeros_like(xgau)
         # get wave solution for these xgau values
         for order_num in set(orders):
-            # find all order_nums that belong to this order
+            # find all order_nums that below to this order
             good = orders == order_num
             # get the xgau for this order
             xgau_order = xgau[good]
@@ -881,8 +858,6 @@ def fit_gaussian_triplets(p , loc):
         dv = np.repeat(np.nan, len(ini_wave_sol))
         # wavelength given in the catalog for the matched line
         wave_catalog = np.zeros(len(ini_wave_sol))
-        # amplitude given in the catolog for the matched lines
-        amp_catalog = np.zeros(len(ini_wave_sol))
         # loop around all lines in ini_wave_sol
         for w_it, wave0 in enumerate(ini_wave_sol):
             # find closest catalog line to the line considered
@@ -892,13 +867,12 @@ def fit_gaussian_triplets(p , loc):
             # check that distance is below threshold
             if np.abs(distv) < cat_guess_dist:
                 wave_catalog[w_it] = wave_ll[id_match]
-                amp_catalog[w_it] = amp_ll[id_match]
                 dv[w_it] = distv
         # ------------------------------------------------------------------
         # Find best trio of lines
         # ------------------------------------------------------------------
         for order_num in set(orders):
-            # find this order's lines
+            # find this orders line
             good = orders == order_num
             # find all usable lines in this order
             good_all = good & (np.isfinite(wave_catalog))
@@ -941,7 +915,7 @@ def fit_gaussian_triplets(p , loc):
             # if we have the minimum number of lines check that we satisfy
             #   the cut_fit_threshold for all good lines and reject outliers
             if bestn >= minimum_number_of_lines:
-                # extrapolate out best fit coefficients over all lines in
+                # extrpolate out best fit coefficients over all lines in
                 #    this order
                 fit_best = np.polyval(best_coeffs, xgau[good])
                 # work out the error in velocity
@@ -975,13 +949,11 @@ def fit_gaussian_triplets(p , loc):
         good = np.isfinite(wave_catalog)
         # apply mask
         wave_catalog = wave_catalog[good]
-        amp_catalog = amp_catalog[good]
         xgau = xgau[good]
         orders = orders[good]
         dv = dv[good]
         ew = ew[good]
         gauss_rms_dev = gauss_rms_dev[good]
-        peak2 = peak2[good]
 
         # ------------------------------------------------------------------
         # Quality check on the total number of lines found
@@ -1056,13 +1028,11 @@ def fit_gaussian_triplets(p , loc):
                 recon0 = recon0[sig_mask]
                 lin_mod_slice = lin_mod_slice[sig_mask]
                 wave_catalog = wave_catalog[sig_mask]
-                amp_catalog = amp_catalog[sig_mask]
                 xgau = xgau[sig_mask]
                 orders = orders[sig_mask]
                 dv = dv[sig_mask]
                 ew = ew[sig_mask]
                 gauss_rms_dev = gauss_rms_dev[sig_mask]
-                peak2 = peak2[sig_mask]
             # Log stats
             sig1 = sig * 1000 / np.sqrt(len(wave_catalog))
             wmsg = '\t{0} | RMS={1:.5f} km/s sig={2:.5f} m/s n={3}'
@@ -1101,7 +1071,6 @@ def fit_gaussian_triplets(p , loc):
 
     # save parameters to loc
     loc['WAVE_CATALOG'] = wave_catalog
-    loc['AMP_CATALOG'] = amp_catalog
     loc['SIG'] = sig
     loc['SIG1'] = sig * 1000 / np.sqrt(len(wave_catalog))
     loc['POLY_WAVE_SOL'] = poly_wave_sol
@@ -1110,9 +1079,6 @@ def fit_gaussian_triplets(p , loc):
     loc['XGAU_T'] = xgau
     loc['ORD_T'] = orders
     loc['GAUSS_RMS_DEV_T'] = gauss_rms_dev
-    loc['DV_T'] = dv
-    loc['EW_T'] = ew
-    loc['PEAK_T'] = peak2
 
     # return loc
     return loc
@@ -1226,9 +1192,9 @@ def generate_resolution_map(p, loc):
             resolution1 = speed_of_light / resolution
             resolution_map[order_num // bin_order, xpos] = resolution1
             # log resolution output
-            wmsg = ('\tOrders {0} - {1}: nlines={2} xpos={3} resolution={4:.5f} km/s '
-                    'R={5:.5f}')
-            wargs = [order_num, order_num + bin_order, np.sum(mask), xpos, resolution,
+            wmsg = ('\tOrder {0}: nlines={1} xpos={2} resolution={3:.5f} km/s '
+                    'R={4:.5f}')
+            wargs = [order_num, np.sum(mask), xpos, resolution,
                      resolution1]
             WLOG('info', p['LOG_OPT'], wmsg.format(*wargs))
         # store criteria
@@ -1401,12 +1367,7 @@ def find_fp_lines_new(p, loc):
     loc = find_fp_lines_new_setup(loc)
     # use spirouRV to get the position of FP peaks from reference file
     loc = spirouRV.CreateDriftFile(p, loc)
-    # remove wide/spurious peaks
-    loc = spirouRV.RemoveWidePeaks(p, loc)
     # check for and remove double-fitted lines
-    # save old position
-    loc['XPEAK_OLD'] = np.copy(loc['XPEAK'])
-    loc['ORDPEAK_OLD'] = np.copy(loc['ORDPEAK'])
     # set up storage for good lines
     ordpeak_k, xpeak_k, ewpeak_k, vrpeak_k, llpeak_k, amppeak_k = \
         [], [], [], [], [], []
@@ -1445,6 +1406,8 @@ def find_fp_lines_new(p, loc):
     loc['VRPEAK'] = np.array(vrpeak_k)
     loc['LLPEAK'] = np.array(llpeak_k)
     loc['AMPPEAK'] = np.array(amppeak_k)
+
+    loc = spirouRV.RemoveWidePeaks(p, loc)
 
     return loc
