+++ INTROOT/bin/cal_WAVE_E2DS_spirou.py
@@ -202,8 +202,7 @@ def main(night_name=None, fpfile=None, hcfiles=None):
         wmsg = 'Calculating FP wave solution'
         WLOG('', p['LOG_OPT'], wmsg)
         # calculate FP wave solution
-        #spirouTHORCA.FPWaveSolution(p, loc, mode=find_lines_mode)
-        spirouTHORCA.FPWaveSolutionNew(p, loc)
+        spirouTHORCA.FPWaveSolution(p, loc, mode=find_lines_mode)
 
         # ------------------------------------------------------------------
         # FP solution plots
