+++ INTROOT/bin/visu_E2DS_spirou.py
@@ -65,10 +65,12 @@ def main(night_name=None, reffile=None):
     # Read image file
     # ----------------------------------------------------------------------
     # read the image data
-    p, fpfitsfilename = spirouStartup.SingleFileSetup(p, filename=p['REFFILE'])
+    gfkwargs = dict(path=p['REDUCED_DIR'], filename=p['REFFILE'])
+    p['REFFILENAME'] = spirouStartup.GetFile(p, **gfkwargs)
+    p.set_source('REFFILENAME', __NAME__ + '/main()')
     # get the fiber type
-    fiber1 = str(p['FIBER'])
-
+    #TODO get the fiber type
+    p['FIBER']='C'
     e2ds, hdr, cmt, nx, ny = spirouImage.ReadImage(p)
     wave = spirouImage.ReadWaveFile(p)
     blaze = spirouImage.ReadBlazeFile(p)
