+++ INTROOT/config/constants_SPIROU_H4RG.py
@@ -735,19 +735,19 @@ ic_errx_min = 0.01  # 0.03
 
 #  Define the wavelength fit polynomial order               - [cal_HC, cal_wave]
 # default = 5
-ic_ll_degr_fit = 4   #5   #4  # 4
+ic_ll_degr_fit = 5   #4  # 4
 
 #  Define the max rms for the wavelength sigma-clip fit     - [cal_HC, cal_wave]
 ic_max_llfit_rms = 3.0
 
 #  Define the fit polynomial order for the Littrow fit      - [cal_HC, cal_wave]
 #      (fit across the orders)
-ic_Littrow_fit_deg_1 = 5  # 5  # 4
-ic_Littrow_fit_deg_2 = 8  # 4
+ic_Littrow_fit_deg_1 = 7  # 5  # 4
+ic_Littrow_fit_deg_2 = 7  # 4
 
 #  Define the littrow cut steps                             - [cal_HC, cal_wave]
 ic_Littrow_cut_step_1 = 250
-ic_Littrow_cut_step_2 = 250    #500
+ic_Littrow_cut_step_2 = 500
 
 #  Define the order to start the Littrow fit from           - [cal_HC, cal_wave]
 #  (ends at ic_hc_n_ord_final)
@@ -759,7 +759,7 @@ ic_Littrow_remove_orders = []
 #  Define the order fit for the Littrow solution            - [cal_HC, cal_wave]
 #      (fit along the orders)
 #TODO needs to be the same as ic_ll_degr_fit
-ic_Littrow_order_fit_deg = 4 #5  # 4
+ic_Littrow_order_fit_deg = 5  # 4
 
 #  Define wavelength free span parameter in find            - [cal_HC, cal_wave]
 #    lines search (used AFTER littrow fit) default = 3
@@ -788,10 +788,10 @@ ic_hc_n_ord_final_2 = 46 # 40    #46
 hc_find_lines_mode = 0
 
 #  Define first order FP solution is calculated from                - [cal_wave]
-ic_fp_n_ord_start =  0   #0   # 9
+ic_fp_n_ord_start =  8   # 9
 
 #  Defines last order FP solution is calculated to                  - [cal_wave]
-ic_fp_n_ord_final = 46    #47   # 45
+ic_fp_n_ord_final =  40   # 45
 
 #  Define the size of region where each line is fitted               -[cal_wave]
 ic_fp_size = 3
@@ -843,6 +843,7 @@ pixel_shift_slope = 0.0
 
 # force reading the wave solution from calibDB
 calib_db_force_wavesol = True
+
 
 # -----------------------------------------------------------------------------
 #   cal_hc/cal_wave parameters
@@ -1086,7 +1087,7 @@ qc_max_signal = 50000
 
 #   Maximum littrow RMS value for cal_hc                    - [cal_HC, cal_wave]
 #       (at x cut points)
-qc_hc_rms_littrow_max = 0.15    #0.3
+qc_hc_rms_littrow_max = 0.1    #0.3
 
 #   Maximum littrow Deviation from wave solution for       - [cal_HC, cal_wave]
 #        cal_wave (at x cut points)
